Class 02 exercise resolution

+++ src/main/kotlin/br/pucpr/authserver/users/UserService.kt
@@ -24,8 +24,11 @@ class UserService(
 
     fun findByIdOrNull(id: Long) = repository.findByIdOrNull(id)
 
-    fun delete(id: Long): Boolean {
+    fun delete(id: Long): Boolean? {
         val user = findByIdOrNull(id) ?: return false
+        if (user.isAdmin() && repository.findByRole("ADMIN").size == 1) {
+            return null
+        }
         repository.delete(user)
         return true
     }
